+++ Scripts/Deploy-Firewall.ps1
@@ -166,18 +166,11 @@ if ($PSBoundParameters.ContainsKey("Confirm"))
 # TODO: It will ask again in same session, seems like set size is not set or Set-ScreenBuffer doesn't work
 Set-ScreenBuffer 3000 @SetScreenBufferParams
 
-# Check all rules which apply to windows services, currently this is used only for debugging purposes
-# NOTE: Most Windows services are not digitally signed and would be reported as untrusted without a -Force switch
-# TODO: A solution is to maintain a list of official Windows services
-# TODO: Rule scripts should Test-Service per rule to avoid loading rules for nonexistent services,
-# a problem though is that multiple scripts may load rules for same service such as svchost,
-# resulting in redundant tests
-if ($Develop)
-{
-	Write-ServiceList $ProjectRoot\Rules -Log | ForEach-Object {
-		Test-Service $_ -Session $SessionInstance -Force | Out-Null
-	}
+# Check all rules which apply to windows services
+Write-ServiceList $ProjectRoot\Rules -Log | ForEach-Object {
+	Test-Service $_ -Session $SessionInstance | Out-Null
 }
+
 Update-Log
 #endregion